pyproject.toml
[project]
name = "ots-sdk"
version = "1.0.0"
description = "Forked Python SDK for the Omnia Timeseries API"
readme = "README.md"
requires-python = ">=3.10,<4.0"

dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "azure-identity >=1.24.0",
    "azure-core",
    "opentelemetry-instrumentation-requests >=0.48b0,<1.0.0",
    "opentelemetry-api >=1.27.0,<2.0.0",
    "cryptography >=43.0.3",
]

[tool.poetry.group.test.dependencies]
pytest = "*"
requests-mock = "*"

[tool.poetry.group.dev.dependencies]

[project.urls]
repository = "https://github.com/asmfstatoil/omnia-timeseries-python"

[build-system]
requires = ["poetry-core>=2.0,<3.0"]
build-backend = "poetry.core.masonry.api"
